refactor: refactor objection plugin
