+++ api/src/services/user/auth.js
@@ -9,6 +9,8 @@ const UNAUTHORIZED = {
 
 const auth = async (instance, next, req, repl) => {
   try {
+    const { User } = instance.models;
+
     const decoded = await req.jwtVerify();
     req.userId = decoded.id;
 
@@ -16,7 +18,7 @@ const auth = async (instance, next, req, repl) => {
       return repl.status(401).send(UNAUTHORIZED);
     }
 
-    const userData = await instance.objection.models.user.query().findOne({
+    const userData = await User.query().findOne({
       id: decoded.id,
     });
